+++ example_project/tests/scrapytests.py
@@ -1,5 +1,9 @@
+# import test classes here
+# test spiders
 from tests.spiders import *
+# item tests
 from tests.items import *
+# stats tests
 from tests.stats import *
 
 # You can define any scrapy settings here